Refactor LSM store for self-containment
* src/store is just a key value store API.
* src/store/lsm implements that API but all the implementation details are it's business.
* consolidated parts(wal, manifest,memtable) of the lsm store that were scattered under one roof.
* Background tasks that do work on the lsm are its domain so they belong under the same roof.
