+++ src/lib.rs
@@ -2,13 +2,9 @@ pub mod config;
 pub mod error;
 pub mod flock;
 pub mod hasher;
-pub mod manifest;
-pub mod memtable;
 pub mod multiwriter;
 pub mod scheduler;
-pub mod sstable;
 pub mod store;
-pub mod wal;
 
 pub use error::Error;
 pub use hasher::Hasher;
