+++ src/store/lsm/mod.rs
@@ -1,0 +1,32 @@
+pub mod compaction;
+pub mod flush;
+pub mod iterator;
+pub mod level;
+pub mod manifest;
+pub mod memtable;
+pub mod metrics;
+pub mod recovery;
+pub mod sstable;
+pub mod state;
+pub mod store;
+pub mod tasks;
+pub mod wal;
+pub mod wal_cleanup;
+
+// Re-export iterator types for use by the main LSM store
+pub use iterator::{HeapEntry, KvIterator, MergeIterator, OwningMemtableIter};
+
+// Re-export level types
+pub use level::{Level, SSTable};
+
+// Re-export memtable types
+pub use memtable::{ActiveMemtable, FrozenMemtable, Memtable, ScanIter, MAX_MEMTABLE_SIZE};
+
+// Re-export state types
+pub use state::{CompactionGuard, FreezeGuard, LsmState};
+
+// Re-export the main LsmStore struct
+pub use store::LsmStore;
+
+// Re-export background tasks
+pub use tasks::{CompactionTask, FlushTask, MetricsTask, WalCleanupTask};
