+++ src/store/lsm/wal/aligned_writer.rs
