+++ src/store/mod.rs
@@ -1,12 +1,4 @@
-pub mod level;
 pub mod lsm;
-pub mod state;
-
-// Background tasks
-pub mod compaction;
-pub mod flush;
-pub mod metrics;
-pub mod wal_cleanup;
 
 use crate::error::Result;
 use std::ops::RangeBounds;
